requirements.txt
Django>=4.2,<5.0
djangorestframework
drf-yasg
celery>=5.2
gunicorn
psycopg2-binary
python-decouple